add new index.html + fix Klausur Status + Session ID
