+++ klausur-service/static/index.html
@@ -22,42 +22,20 @@
 <div class="container-fluid">
     <div class="row">
         <div class="col">
-            <form class="studentDataForm">
+            <form id="KlausurIDForm">
                 <div class="form-floating mb-3">
-                    <input class="form-control" id="studentenNameInput" placeholder="Name">
-                    <label for="studentenNameInput">Name</label>
+                    <input class="form-control" id="KlausurIDInput" placeholder="Klausur ID">
+                    <label for="KlausurIDInput">Klausur ID</label>
                 </div>
-                <div class="form-floating">
-                    <input class="form-control" id="studentenMatrikelnummerInput" placeholder="Matrikelnummer">
-                    <label for="studentenMatrikelnummerInput">Matrikelnummer</label>
-                </div>
+                <button type="submit">Send</button>
             </form>
-            <div class="form-check form-switch">
-                <input class="form-check-input" type="checkbox" id="SwitchCheckNameNummer">
-                <label class="form-check-label" for="SwitchCheckNameNummer">Name und Matrikelnummer speichern und blockieren</label>
-            </div>
-        </div>
-        <div class="col studentDataStatus">
-            <span id="studentDataStatus"></span>
         </div>
     </div>
-    <div class="row">
-        <div class="col klausurstatus">
-            <h1>Klausurstatus:</h1> <span id="klausurstatus"></span>
-        </div>
+    <div class="col studentDataStatus">
+        <span id="KlausurIDDataStatus"></span>
     </div>
 </div>
 
-<div class="container-fluid" style="height: 5000px">
-    <div class="sticky-top">
-        <div class="position-absolute end-0">Zeit: <span id="timer-text"></span></div>
-    </div>
-    <form id="klausurBodyForm">
-        <div id="klausurbody">
-        </div>
-    </form>
-</div>
 <script src="js/indexFunktions.js"></script>
-<script src="js/timer.js"></script>
 </body>
 </html>
